refactor: mobile에서만 보여지는 tabBar - boardMobileTab으로 이동
